% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeu.R
\name{a_creuser}
\alias{a_creuser}
\title{creuser les cases qui contiennent 0
Liste des case sans bombes dans une grille de démineur}
\usage{
a_creuser(grille)
}
\arguments{
\item{grille}{Une grille de démineur}
}
\value{
Un vecteur constitué l'identifiant des cases contenant des bombes dans la grille
}
\description{
creuser les cases qui contiennent 0
Liste des case sans bombes dans une grille de démineur
}
