% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeu.R
\name{creuser}
\alias{creuser}
\title{la fonction qui creuse une case d'une grille de démineur}
\usage{
creuser(plateau, grille, case)
}
\arguments{
\item{plateau}{Un plateau de jeu de la même taille que la grille}

\item{grille}{Une grille de démineur}

\item{case}{L'identifiant de la case à creuser}
}
\value{
Le plateau de jeu creusé à la case indiqué si ce n'est pas une bombe. La grille de jeu si la case creusée contient une bombe.
}
\description{
la fonction qui creuse une case d'une grille de démineur
}
\examples{
Grille <- grille(5,5)
Plateau <- matrix(1:(5*5), nrow = 5, ncol = 5)
creuser(Plateau, Grille, 4)
creuser(Plateau, Grille, 16)
}
