% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeu.R
\name{gagne}
\alias{gagne}
\title{Afficher la victoire ou l'échec}
\usage{
gagne(creuse, acreuser, plateau)
}
\arguments{
\item{creuse}{(Vecteur) L'identifiant des cases déjà creusées.}

\item{acreuser}{(Vecteur) L'identifiant de toutes les cases sans bombes.}

\item{plateau}{(Matrice) Un plateau de jeu.}
}
\value{
Une chaîne de caractère indiquant si la partie est gagnée,
ou perdue.
}
\description{
Cette fonction indique si la partie est gagnée ou perdue.
}
\examples{
G <- grille(8,9)
P <- matrix(1:(8*9), nrow = 8, ncol = 9)
acreuser <- a_creuser(G)
creuse <- c(4,2,15,6,9,20)
gagne(creuse, acreuser, P)
}
