% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grille.R
\name{grille}
\alias{grille}
\title{Grille de démineur}
\usage{
grille(L, C)
}
\arguments{
\item{L}{(Entier) Le nombre de lignes.}

\item{C}{(Entier) Le nombre de colonnes.}
}
\value{
Une matrice à L lignes, C colonnes
}
\description{
Cette fonction créé une matrice d'un nombre de lignes et de
colonnes données contenant des bombes et le nombre de bombes autour de
chaque case vide.
}
\examples{
grille(5,6)
grille(9,7)
}
