% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grille.R
\name{nbr_bombe}
\alias{nbr_bombe}
\title{Compte le nombre de bombe dans une grille de démineur}
\usage{
nbr_bombe(grille)
}
\arguments{
\item{grille}{Une grille de démineur}
}
\value{
Le nombre de bombe dans la grille de démineur et leur position
}
\description{
Compte le nombre de bombe dans une grille de démineur
}
\examples{
Grille <- grille(5,5)
nbr_bombe(Grille)
}
