% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{runapp}
\alias{runapp}
\title{Lancement de l'application}
\usage{
runapp()
}
\value{
Application shiny "Démineur"
}
\description{
Cette fonction permet de lancer l'application shiny du package
"demineurER".
}
\examples{
runapp()
}
