% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeu.R
\name{a_creuser}
\alias{a_creuser}
\title{Case à creuser dans une grille démineur}
\usage{
a_creuser(grille)
}
\arguments{
\item{grille}{Grille de démineur (matrice contenant des bombes)}
}
\value{
Un vecteur constitué l'identifiant des cases contenant des bombes
dans la grille
}
\description{
Liste des identifiants de toutes les cases sans bombes dans une grille de
démineur.
}
\examples{
g1 <- grille(4,6)
a_creuser(g1)
g2 <- grille(10,10)
a_creuser(g2)
}
