% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeu.R
\name{gagne}
\alias{gagne}
\title{Afficher la victoire ou l'échec}
\usage{
gagne(creuse, acreuser, plateau)
}
\arguments{
\item{creuse}{Identifiant des cases déjà creusées.}

\item{acreuser}{Identifiant de toutes les cases sans bombes.}
}
\value{
Une chaîne de caractère indiquant si la partie est gagnée,
ou perdue.
}
\description{
Cette fonction indique si la partie est gagnée ou perdue.
}
\examples{
G <- grille(8,9)
acreuser <- a_creuser(G)
creuse <- c(4,2,15,6,9,20)
gagne(creuse, acreuser)
}
