% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeu.R
\name{gagne}
\alias{gagne}
\title{Etat de la partie de démineur}
\usage{
gagne(creuse, acreuser)
}
\arguments{
\item{creuse}{Identifiant des cases creusées.}

\item{acreuser}{Identifiant des cases à creuser pour gagner la partie.}
}
\value{
Une chaîne de caractère indiquant si la partie est gagnée, perdue ou toujours en cours.
}
\description{
Etat de la partie de démineur
}
