% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grille.R
\name{grille}
\alias{grille}
\title{Créé une grille de démineur}
\usage{
grille(L, C)
}
\arguments{
\item{L}{Un nombre de ligne}

\item{C}{Un nombre de colonne}
}
\value{
Une matrice à L linges, C colonnes
}
\description{
Créé une grille de démineur
}
