% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grille.R
\name{grille}
\alias{grille}
\title{Créér une grille du jeu démineur}
\usage{
grille(L, C)
}
\arguments{
\item{L}{le nombre de ligne}

\item{C}{le nombre de colonne}
}
\value{
Une matrice à L lignes, C colonnes
}
\description{
Créér une grille du jeu démineur
}
