% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grille.R
\name{nbr_bombe}
\alias{nbr_bombe}
\title{Nombre de bombe}
\usage{
nbr_bombe(grille)
}
\arguments{
\item{grille}{Grille de démineur}
}
\value{
Le nombre de bombe dans la grille de démineur.
}
\description{
Cette fonction compte le nombre de bombe dans une grille de
démineur créée avec la fonction **grille**
}
\examples{
Grille <- grille(5,5)
nbr_bombe(Grille)
}
